Update wcwidth.c with Unicode 9.0.0 data
In order to maintain compatibility with screen[1], update
wcwidth functions to use Unicode 9.0.0 character database.

Updated intervals extracted from output of [2] from ucd files[3].

The comments at the head of [2] state that the output is unrestricted.
Therefore it is not subject to the GPL as applies to the script itself.

[1] See: https://savannah.gnu.org/bugs/?func=detailitem&item_id=50044
[2] https://raw.githubusercontent.com/GNOME/glib/37d4c2941bd0326b8b6e6bb22c81bd424fcc040b/glib/gen-unicode-tables.pl
[3] http://www.unicode.org/Public/9.0.0/ucd/